Move the animated background to separate file
