+++ frontend/src/components/password.module.css
@@ -1,20 +1,3 @@
-.backgroundContainer {
-	position: fixed;
-	top: 0;
-	left: 0;
-	width: 100%;
-	height: 100%;
-	background: linear-gradient(135deg, #1a1a1a 0%, #2d2d2d 100%);
-	z-index: -1;
-	overflow: hidden;
-}
-
-.backgroundSvg {
-	position: absolute;
-	top: 0;
-	left: 0;
-}
-
 #container {
 	display: flex;
 	justify-content: center;
